Cleanup fallthrough, printf, and unused variable level-3 warnings

+++ src/cpu/core_full/string.h
@@ -241,7 +241,9 @@
 		}
 		break;
 	default:
-		LOG(LOG_CPU,LOG_ERROR)("Unhandled string %d entry %X",inst.code.op,inst.entry);
+		LOG(LOG_CPU, LOG_ERROR)
+		("Unhandled string %d entry %X", inst.code.op,
+		 static_cast<uint32_t>(inst.entry));
 	}
 	/* Clean up after certain amount of instructions */
 	reg_esi&=(~add_mask);
